Unit test every function

+++ tests/testthat/utils.R
@@ -1,11 +1,13 @@
 testwd = function(x){
   dir = paste0("RUN-", x)
+  if(file.exists(file.path("..", dir))) return()
   if(!file.exists(dir)) dir.create(dir)
   setwd(dir)
 }
 
 testrm = function(x){
   dir = paste0("RUN-", x)
+  if(!file.exists(file.path("..", dir))) return()
   setwd("..")
   unlink(dir, recursive = T)
 }
